lab3 - error handling

+++ jaxws/server/src/main/java/ru/taravkov/ifmo/webservices/ws/service/CarWebService.java
@@ -6,6 +6,8 @@ import ru.taravkov.ifmo.webservices.dao.CarDaoImpl;
 import ru.taravkov.ifmo.webservices.entity.Car;
 import ru.taravkov.ifmo.webservices.entity.VehicleClass;
 import ru.taravkov.ifmo.webservices.entity.Color;
+import ru.taravkov.ifmo.webservices.ws.exception.ServiceException;
+import ru.taravkov.ifmo.webservices.ws.exception.ServiceException.Detail;
 
 import javax.jws.WebMethod;
 import javax.jws.WebParam;
@@ -42,6 +44,21 @@ public class CarWebService {
                           @WebParam(name = "color") Color color,
                           @WebParam(name = "vehicleClass") VehicleClass vehicleClass,
                           @WebParam(name = "rightHand") Boolean rightHand) {
+        if (make == null) {
+            throw new ServiceException("Unable to create car record", new Detail("Make can not be null"));
+        }
+        if (model == null) {
+            throw new ServiceException("Unable to create car record", new Detail("Model can not be null"));
+        }
+        if (color == null) {
+            throw new ServiceException("Unable to create car record", new Detail("Color can not be null"));
+        }
+        if (vehicleClass == null) {
+            throw new ServiceException("Unable to create car record", new Detail("Vehicle class can not be null"));
+        }
+        if (rightHand == null) {
+            throw new ServiceException("Unable to create car record", new Detail("Right hand attribute can not be null"));
+        }
         return carDao.create(make, model, color, vehicleClass, rightHand);
     }
 
@@ -55,6 +72,24 @@ public class CarWebService {
                            @WebParam(name = "color") Color color,
                            @WebParam(name = "vehicleClass") VehicleClass vehicleClass,
                            @WebParam(name = "rightHand") Boolean rightHand) {
+        if (id == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Id can not be null"));
+        }
+        if (make == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Make can not be null"));
+        }
+        if (model == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Model can not be null"));
+        }
+        if (color == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Color can not be null"));
+        }
+        if (vehicleClass == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Vehicle class can not be null"));
+        }
+        if (rightHand == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Right hand attribute can not be null"));
+        }
         return carDao.update(id, make, model, color, vehicleClass, rightHand) ? OpStatus.SUCCESS : OpStatus.FAILURE;
     }
 
@@ -63,6 +98,9 @@ public class CarWebService {
      */
     @WebMethod(operationName = "deleteCar")
     public OpStatus delete(@WebParam(name = "id") Long id) {
+        if (id == null) {
+            throw new ServiceException("Unable to update car record", new Detail("Id can not be null"));
+        }
         return carDao.delete(id) ? OpStatus.SUCCESS : OpStatus.FAILURE;
     }
 }
